requirements.txt
streamlit
openai
langchain
PyPDF2
python-dotenv
sentence-transformers
faiss-cpu
